+++ service/src/main/java/com/lau/cargo/entity/Booking.java
@@ -1,0 +1,28 @@
+package com.lau.cargo.entity;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.Date;
+@Getter
+@Setter
+@Entity
+public class Booking {
+    @Id
+    @GeneratedValue(strategy=GenerationType.AUTO)
+    private Long bookingId;
+    private Date startDate;
+    private Date endDate;
+    private Double bookingPrice;
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "userId", referencedColumnName = "userId")
+    private User user;
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "carId", referencedColumnName = "carId")
+    private Car car;
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "addressId", referencedColumnName = "addressId")
+    private Address address;
+
+
+}
