+++ service/src/main/java/com/lau/cargo/entity/Car.java
@@ -1,0 +1,36 @@
+package com.lau.cargo.entity;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.util.List;
+
+
+@Getter
+@Setter
+@Entity
+public class Car {
+    @Id
+    @GeneratedValue(strategy=GenerationType.AUTO)
+    private Long carId;
+    private String make;
+    private String model;
+    private Integer year;
+    private String color;
+    private String availableFrom;
+    private String availableTo;
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "userId", referencedColumnName = "userId")
+    private User user;
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "addressId", referencedColumnName = "addressId")
+    private Address address;
+    @OneToMany(mappedBy = "car", fetch = FetchType.LAZY,
+            cascade = CascadeType.ALL)
+    private List<Booking> booking;
+    @OneToMany(mappedBy = "car", fetch = FetchType.LAZY,
+            cascade = CascadeType.ALL)
+    private List<Review> review;
+}
+
+
