+++ service/src/main/java/com/lau/cargo/entity/Review.java
@@ -1,0 +1,23 @@
+package com.lau.cargo.entity;
+
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.Setter;
+
+@Entity
+@Setter
+@Getter
+public class Review {
+    @Id
+    @GeneratedValue(strategy= GenerationType.AUTO)
+    private Long reviewId;
+    private String reviewText;
+    private Double rating;
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "userId", referencedColumnName = "userId")
+    private User user;
+    @ManyToOne(cascade = CascadeType.ALL)
+    @JoinColumn(name = "carId", referencedColumnName = "carId")
+    private Car car;
+
+}
